server/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
Automat==22.10.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
constantly==15.1.0
crochet==2.0.0
cryptography==40.0.2
cssselect==1.2.0
dnspython==2.3.0
filelock==3.12.0
frozenlist==1.3.3
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
itemadapter==0.8.0
itemloaders==1.1.0
jmespath==1.0.1
lxml==4.9.2
multidict==6.0.4
openai==0.27.8
packaging==23.1
parsel==1.8.1
Protego==0.2.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyDispatcher==2.0.7
pymongo==4.3.3
pyOpenSSL==23.1.1
python-dotenv==1.0.0
queuelib==1.6.2
requests==2.31.0
requests-file==1.5.1
Scrapy==2.9.0
service-identity==21.1.0
six==1.16.0
tldextract==3.4.4
tqdm==4.65.0
Twisted==22.10.0
typing-extensions==4.5.0
urllib3==2.0.2
w3lib==2.1.1
wrapt==1.15.0
yarl==1.9.2
zope.interface==6.0
